t145: split RestController.php (6,413 lines) into 9 domain controllers
Extract handler methods from the monolithic RestController into separate
domain-specific controller classes, each with its own register_routes().
RestController is now a thin orchestrator (~500 lines) that delegates to:

- SessionController    — sessions, jobs, process, run, site-builder
- SettingsController   — settings, providers, budget, usage, roles, alerts
- MemoryController     — memories
- SkillController      — skills
- AutomationController — automations, event-automations, logs, triggers
- KnowledgeController  — knowledge collections, sources, search, stats
- ToolController       — custom-tools, tool-profiles, abilities
- ChangesController    — changes, modified-plugins, download
- AgentController      — agents, conversation-templates

Shared permission methods extracted to PermissionTrait (used by all controllers).
No route URLs changed. PHPCS passes with zero violations.

Closes #729
